created prefab "Visual Form"
has components MeshRenderer and MeshFilter, this is used to add to a game object and be its visual representation. A game object can have multiple visualforms that make up its appearance. therefore a game object can also have multiple visual form prefabs attached. Set the transform of the visualform object and it will offset it from the parent gameobject